bug fix
Former-commit-id: d263a56f1aee405156f9200b72f8c3a2fa9bf5a7
Former-commit-id: 2f2604f6f4809bc2837f68933368dd5dc371ab9f
Former-commit-id: 7b5b619c581c7dbf2ea42618a7e76966b609c65a
Former-commit-id: f5969cf6f76158670d66ea2328dcb1e3cb0ef972 [formerly 0c00f3741f5227afb66f2abf5aad51643f4a3a13]
Former-commit-id: 4af37a7068f8e93c2932cc1959422d6ca149d523

+++ src/tamm/ops.hpp
@@ -385,7 +385,7 @@ inline void block_add(T* dbuf, const std::vector<size_t>& ddims,
         auto idx = [](const auto& index_vec, const auto& dims_vec) {
             size_t ret = 0, ld = 1;
             EXPECTS(index_vec.size() == dims_vec.size());
-            for(size_t i = index_vec.size(); i >= 0; i--) {
+            for(int i = index_vec.size(); i >= 0; i--) {
                 ret += ld * index_vec[i];
                 ld *= dims_vec[i];
             }
